+++ packages/radix-ui-themes/src/components/text-area.props.ts
@@ -2,11 +2,11 @@ import { colorProp } from '../helpers';
 import { PropDef } from '../helpers';
 
 const sizes = ['1', '2', '3'] as const;
-const variants = ['solid', 'classic', 'soft'] as const;
+const variants = ['surface', 'classic', 'soft'] as const;
 
 const textAreaPropDefs = {
   size: { type: 'enum', values: sizes, default: '2', responsive: true },
-  variant: { type: 'enum', values: variants, default: 'solid' },
+  variant: { type: 'enum', values: variants, default: 'surface' },
   color: colorProp,
 } satisfies {
   size: PropDef<(typeof sizes)[number]>;